refactor(viewer): extract useViewerTimeline from GeoViewer
Decomposition 2/4, content-neutral: the union time axis, step/clip
selection with URL-seeded relocation, the time-link policy, per-side
resolution, hover/display projections, failure-mark projection, and
offset tags move into a hook. The GetMap failure log stays with
GeoViewer (bound to source lifecycles) and feeds the hook its marks;
URL seeds enter as an explicit "initial" param. Tests untouched.
